requirements.txt
# Vehicle Communication
pyserial==3.5

# Configuration Management
pyyaml>=6.0

# AI Backends
anthropic>=0.18.0
openai>=1.12.0

# Testing
pytest>=7.4.0
hypothesis>=6.92.0  # Property-based testing
